// src/@core/purchase-data-types.ts
export type DetailPurchaseOrderType = {
  product_id: number
  product_name: string
  barcode: string
  qty_request: number
  qty_received: number
  qty_over: number
  price: number
  percent_discount: number
  total_price: number
  unit_name: string
  initial: string
  branch_name: string
}

export type PurchaseOrderType = {
  id: number
  number: string
  ref?: string
  date: Date
  due_date: Date
  term?: string
  description?: string | null
  total_qty_request: number
  total_qty_received: number
  total_qty_over: number
  total_price: number
  total_discount: number
  total_tax: number
  total_charge: number
  grand_total: number
  qty: number
  vendor_name: string
  vendor_id: number
  warehouse_id: number
  branch_name: string
  branch_id: number
  detail?: DetailPurchaseOrderType[]
  status_id: number
  status: string
}